allow KEEP_TARGET_DIRECTORY

+++ entrypoint.sh
@@ -103,13 +103,16 @@ mv "$CLONE_DIR/.git" "$TEMP_DIR/.git"
 # $TARGET_DIRECTORY is '' by default
 ABSOLUTE_TARGET_DIRECTORY="$CLONE_DIR/$TARGET_DIRECTORY/"
 
-echo "[+] Deleting $ABSOLUTE_TARGET_DIRECTORY"
-rm -rf "$ABSOLUTE_TARGET_DIRECTORY"
+if [[ ! $KEEP_TARGET_DIRECTORY ]]
+then
+	echo "[+] Deleting $ABSOLUTE_TARGET_DIRECTORY"
+	rm -rf "$ABSOLUTE_TARGET_DIRECTORY"
+fi
 
-echo "[+] Creating (now empty) $ABSOLUTE_TARGET_DIRECTORY"
+echo "[+] Ensuring $ABSOLUTE_TARGET_DIRECTORY"
 mkdir -p "$ABSOLUTE_TARGET_DIRECTORY"
 
-echo "[+] Listing contents of target (should be empty) $ABSOLUTE_TARGET_DIRECTORY"
+echo "[+] Listing contents of target $ABSOLUTE_TARGET_DIRECTORY"
 ls -al "$ABSOLUTE_TARGET_DIRECTORY"
 
 echo "[+] Listing Current Directory Location"
